Use Django database router API

+++ src/reactpy_django/utils.py
@@ -24,7 +24,6 @@ from django.views import View
 
 from reactpy_django.exceptions import ComponentDoesNotExistError, ComponentParamError
 
-
 _logger = logging.getLogger(__name__)
 _component_tag = r"(?P<tag>component)"
 _component_path = r"(?P<path>\"[^\"'\s]+\"|'[^\"'\s]+')"
@@ -334,7 +333,6 @@ def db_cleanup(immediate: bool = False):
     This function may be expanded in the future to include additional cleanup tasks."""
     from .config import (
         REACTPY_CACHE,
-        REACTPY_DATABASE,
         REACTPY_DEBUG_MODE,
         REACTPY_RECONNECT_MAX,
     )
@@ -351,7 +349,7 @@ def db_cleanup(immediate: bool = False):
     expires_by: datetime = timezone.now() - timedelta(seconds=REACTPY_RECONNECT_MAX)
 
     # Component params exist in the DB, but we don't know when they were last cleaned
-    if not cleaned_at_str and ComponentSession.objects.using(REACTPY_DATABASE).all():
+    if not cleaned_at_str and ComponentSession.objects.all():
         _logger.warning(
             "ReactPy has detected component sessions in the database, "
             "but no timestamp was found in cache. This may indicate that "
@@ -361,9 +359,7 @@ def db_cleanup(immediate: bool = False):
     # Delete expired component parameters
     # Use timestamps in cache (`cleaned_at_str`) as a no-dependency rate limiter
     if immediate or not cleaned_at_str or timezone.now() >= clean_needed_by:
-        ComponentSession.objects.using(REACTPY_DATABASE).filter(
-            last_accessed__lte=expires_by
-        ).delete()
+        ComponentSession.objects.filter(last_accessed__lte=expires_by).delete()
         caches[REACTPY_CACHE].set(cache_key, now_str, timeout=None)
 
     # Check if cleaning took abnormally long
